requirements.txt
openai>=1.0.0,<2.0.0
py_trees>=2.3.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
